v_0.75_Demo
- 5층 확장시 장애물 Show_Sprite가 안되고 있던 부분 수정
- 이벤트 박스 폰트색 수정
- 비밀방 포탈 위치를 오른쪽으로 한칸 옮김
- 장애물이 우물이미지로 되있던 부분 수정
- 특수시설을 4층부터는 매 층마다 추가되도록 변경(이럼 확장할 이유도 생김 / 특수시설에 넣을게 생각보다 많음(이벤트 시설도 그렇고 전체에 단 한개만 존재해야 하는 퍼실리티는 사실 특수시설로 하는게 맞기도 하고)
    - 일단 테스트는 해봐야함. 5층 확장시 시설 3이 생기는지
    - 특수시설 - 치료 시설 추가
        - 체력 보너스 15
        - 부상회복비용 50% 감소
- 장애물 제거비용 - 마나 코스트 절반으로 줄임
- 1층에 3*3 장애물을 없애고 1*1 장애물을 2개 추가함
- 배틀필드를 1,2,3 통일 / 0,4,5 통일로 변경함(색이 안맞음)
- 솔져 1,2,3중에 1은 퍼실리티도 탐색하도록 변경(20일차 난이도가 너무 높음)
- 몬스터 소환비용 소폭 너프(약 10%가량씩)
- 20일에 대화가 나오면 데모클리어 위에 뜨는 현상 수정
- 유닛 이벤트 룸 추가
- 유닛 이벤트 테스트용 - 슬라임과 히로인 첫번째 대화 추가
- 5층확장비용 1000,500으로 감소

+++ DeongeonManagement/Assets/Script/Facility/Non_Inherit/Orb.cs
@@ -204,7 +204,7 @@ public class Orb : MonoBehaviour
                 var green = $"{UserData.Instance.LocaleText_Tooltip("Orb_Green")}\n" +
                     $"({UserData.Instance.LocaleText_Tooltip("Orb_Green_Detail")})";
                 ui.SetText(green, () => Confirm(ap: 2, mana: 750, gold: 250));
-                ui.SetMode_Calculation((int)Define.DungeonRank.F, "750", "250", "2");
+                ui.SetMode_Calculation(Define.DungeonRank.F, "750", "250", "2");
                 break;
 
 
@@ -212,7 +212,7 @@ public class Orb : MonoBehaviour
                 var yellow = $"{UserData.Instance.LocaleText_Tooltip("Orb_Yellow")}\n" +
                     $"({UserData.Instance.LocaleText_Tooltip("Orb_Yellow_Detail")})";
                 ui.SetText(yellow, () => Confirm(ap: 2, mana: 750, gold: 250));
-                ui.SetMode_Calculation((int)Define.DungeonRank.F, "750", "250", "2");
+                ui.SetMode_Calculation(Define.DungeonRank.F, "750", "250", "2");
                 break;
 
 
@@ -220,7 +220,7 @@ public class Orb : MonoBehaviour
                 var blue = $"{UserData.Instance.LocaleText_Tooltip("Orb_Blue")}\n" +
                     $"({UserData.Instance.LocaleText_Tooltip("Orb_Blue_Detail")})";
                 ui.SetText(blue, () => Confirm(mana: 3000, gold: 0, ap: 3));
-                ui.SetMode_Calculation((int)Define.DungeonRank.D, "3000", "0", "3");
+                ui.SetMode_Calculation(Define.DungeonRank.D, "3000", "0", "3");
                 break;
 
 
@@ -228,7 +228,7 @@ public class Orb : MonoBehaviour
                 var red = $"{UserData.Instance.LocaleText_Tooltip("Orb_Red")}\n" +
                     $"({UserData.Instance.LocaleText_Tooltip("Orb_Red_Detail")})";
                 ui.SetText(red, () => Confirm(mana: 1000, gold: 2000, ap: 4));
-                ui.SetMode_Calculation((int)Define.DungeonRank.C, "1000", "2000", "4");
+                ui.SetMode_Calculation(Define.DungeonRank.C, "1000", "2000", "4");
                 break;
         }
     }
